add access checks on various controllers

+++ src/main/java/white/monster/energy/adventurebackend/booking/BookingController.java
@@ -1,21 +1,24 @@
 package white.monster.energy.adventurebackend.booking;
 
+import jakarta.servlet.http.HttpServletRequest;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.format.annotation.DateTimeFormat;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import white.monster.energy.adventurebackend.profile.ProfileType;
 
 import java.net.URI;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.stream.*;
 
+@CrossOrigin
 @RestController
 @RequestMapping("/api/bookings")
-@CrossOrigin
 public class BookingController {
 
     private final BookingService service;
@@ -34,8 +37,11 @@ public class BookingController {
             @RequestParam(required = false)
             @DateTimeFormat(iso = DateTimeFormat.ISO.DATE_TIME) LocalDateTime from,
             @RequestParam(required = false)
-            @DateTimeFormat(iso = DateTimeFormat.ISO.DATE_TIME) LocalDateTime to
+            @DateTimeFormat(iso = DateTimeFormat.ISO.DATE_TIME) LocalDateTime to,
+            HttpServletRequest request
     ) {
+//        if (ProfileType.ADMIN.verifyAccessLevel(request)) return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+
         Pageable pageable = PageRequest.of(page, size);
 
         // find bookings that match the search
@@ -63,7 +69,9 @@ public class BookingController {
 
     // Find a booking by its ID and show it.
     @GetMapping("/{id}")
-    public ResponseEntity<BookingDto> get(@PathVariable int id) {
+    public ResponseEntity<BookingDto> get(@PathVariable int id, HttpServletRequest request) {
+        if (ProfileType.ADMIN.verifyAccessLevel(request)) return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+
         Booking b = service.getById(id);
         return (b == null) ? ResponseEntity.notFound().build()
                 : ResponseEntity.ok(BookingDto.from(b));
@@ -79,7 +87,8 @@ public class BookingController {
 
     // Change a booking that already exists.
     @PatchMapping("/{id}")
-    public ResponseEntity<BookingDto> update(@PathVariable int id, @RequestBody BookingDto dto) {
+    public ResponseEntity<BookingDto> update(@PathVariable int id, @RequestBody BookingDto dto, HttpServletRequest request) {
+        if (ProfileType.ADMIN.verifyAccessLevel(request)) return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
         Booking updated = service.update(id, dto);
         return (updated == null) ? ResponseEntity.notFound().build()
                 : ResponseEntity.ok(BookingDto.from(updated));
@@ -87,7 +96,8 @@ public class BookingController {
 
     // Delete a booking by its ID.
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> delete(@PathVariable int id) {
+    public ResponseEntity<Void> delete(@PathVariable int id, HttpServletRequest request) {
+        if (ProfileType.ADMIN.verifyAccessLevel(request)) return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
         boolean ok = service.delete(id);
         return ok ? ResponseEntity.noContent().build() : ResponseEntity.notFound().build();
     }
@@ -95,7 +105,9 @@ public class BookingController {
     @PostMapping("/assign-employee")
     public ResponseEntity<?> assignEmployee(
             @RequestParam int bookingId,
-            @RequestParam int employeeId) {
+            @RequestParam int employeeId,
+            HttpServletRequest request) {
+        if (ProfileType.ADMIN.verifyAccessLevel(request)) return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
         try {
             Booking updated = service.assignEmployee(bookingId, employeeId);
             return ResponseEntity.ok(BookingDto.from(updated));
@@ -105,7 +117,8 @@ public class BookingController {
     }
 
     @GetMapping("/assigned-to/{employeeId}")
-    public ResponseEntity<List<BookingDto>> getBookingsForEmployee(@PathVariable int employeeId) {
+    public ResponseEntity<List<BookingDto>> getBookingsForEmployee(@PathVariable int employeeId, HttpServletRequest request) {
+        if (ProfileType.OPERATOR.verifyAccessLevel(request)) return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
         List<Booking> bookings = service.getBookingsForEmployee(employeeId);
         List<BookingDto> dtos = bookings.stream()
                 .map(BookingDto::from)
